feat(forms): Refactor forms to allow both context managers and plain validation
Rename form.py to forms.py

Refactor Form() function to just validate the form against the pydantic model and return the model. This is no longer a context manager.

Create FormManager() which is an AsyncContextManager. Can be used to validate the form and stay within the context of the request.form() to allow working with files.

BREAKING CHANGE: